Renames ProfileCTAScreen to ProfilePurchaseScreen - for improved consistency with other profile screens; Refactores uses of old name to new one.
