+++ cardstack/src/screens/Profile/index.ts
@@ -1,4 +1,4 @@
 export { default as ProfileNameScreen } from './ProfileNameScreen/ProfileNameScreen';
 export { default as ProfileSlugScreen } from './ProfileSlugScreen/ProfileSlugScreen';
-export { default as ProfilePurchaseCTA } from './PurchaseCTAScreen/PurchaseCTAScreen';
+export { default as ProfilePurchaseScreen } from './ProfilePurchaseScreen/ProfilePurchaseScreen';
 export { default as ProfileChargeExplanationScreen } from './ProfileChargeExplanationScreen/ProfileChargeExplanationScreen';
